feat-add-global-style-for-header-footer-and-screen-wide-lines-NMO-714

+++ dockerized-norlab-images/core-images/dn-project/dn_callback_execute_compose_post.bash
@@ -18,13 +18,11 @@
 #
 # =================================================================================================
 function dn::callback_execute_compose_post() {
-  set -e
-
   n2st::print_formated_script_header "dn_callback_execute_compose_post.bash" "${MSG_LINE_CHAR_UTIL}"
 
-  test -n "${DN_PATH:?err}" || n2st::print_msg_error_and_exit "Env variable DN_PATH need to be set and non-empty."
+  test -n "${DN_PATH:?err}" || { n2st::print_msg_error "Env variable DN_PATH need to be set and non-empty." && return 1 ; }
   test -d "${DN_PATH}/utilities/tmp/dockerized-norlab-project-mock" \
-    || n2st::print_msg_error_and_exit "The directory ${DN_PATH}/utilities/tmp/dockerized-norlab-project-mock is unreachable"
+    || { n2st::print_msg_error "The directory ${DN_PATH}/utilities/tmp/dockerized-norlab-project-mock is unreachable" && return 1 ; }
 
   # Delete git cloned repo
   rm -rf "${DN_PATH}/utilities/tmp/dockerized-norlab-project-mock"
@@ -32,14 +30,13 @@ function dn::callback_execute_compose_post() {
   # Setup placeholder
   mkdir "${DN_PATH}/utilities/tmp/dockerized-norlab-project-mock"
 
-  # ....Saniti check...............................................................................
-  test -d "${DN_PATH}/utilities" || n2st::print_msg_error_and_exit "The directory ${DN_PATH}/utilities is unreachable"
+  # ....Sanity check...............................................................................
+  test -d "${DN_PATH}/utilities" || { n2st::print_msg_error "The directory ${DN_PATH}/utilities is unreachable" && return 1 ; }
   test ! -d "${DN_PATH}/utilities/tmp/dockerized-norlab-project-mock/.git" \
   || { \
     tree -a -L 2 "${DN_PATH}/utilities" &&
-    n2st::print_msg_error_and_exit "The directory ${DN_PATH}/utilities is unreachable" ;
+    n2st::print_msg_error "The directory ${DN_PATH}/utilities is unreachable" && return 1 ;
     }
-
 
   n2st::print_formated_script_footer "dn_callback_execute_compose_post.bash" "${MSG_LINE_CHAR_UTIL}"
 }
